src/python_container/requirements.txt
python-binance
pandas
sqlalchemy
psycopg2-binary
tqdm
websocket